### 2018-07-24 #839 @leeper
@davidnsousa Okay, thanks for those details. The term "social behavior" is unfortunately a bit ambiguous for a multidisciplinary audience as it could also imply sociology, anthropology, etc. Can you update the paper and README to reflect your meaning, perhaps with an example?

And yes, please add a LICENSE/LICENSE.txt/LICENSE.md file (and add it to `.Rbuildignore`).
