### 2018-07-24 #839 @davidnsousa
@leeper thank you. Added example to README and paper. Is it more clear now? Also added LICENSE (and added LICENSE to `.Rbuildignore`)
