### 2018-07-24 c67c5ff
add CRAN status badge
