+++ README.Rmd
@@ -9,6 +9,7 @@ knitr::opts_chunk$set(echo = TRUE)
 library(bor)
 ```
 
+[![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/bor)](https://cran.r-project.org/package=bor)
 [![Build Status](https://travis-ci.org/davidnsousa/bor.svg?branch=master)](https://travis-ci.org/davidnsousa/bor) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1317543.svg)](https://doi.org/10.5281/zenodo.1317543)
 
 # bor
