### 2018-07-31 7d743a7
added rdoc badge
