### 2018-08-21 #839 @ViniciusBRodrigues
Hi,
I made an initial review and the submission looks great. It's everything good with the review criteria. One think that they need to do is complete the **Community guidelines:**

> There should be clear guidelines for third-parties wishing to:
>
> - Contribute to the software
> - Report issues or problems with the software
> - Seek support

The minor suggestion is break the follow line in the paper:
 `## specify the set of behaviors recorded and that will be used in bset argument`
