### 2018-08-22 #839 @davidnsousa
@ViniciusBRodrigues should I do this changes now? Where do I complete the community guidelines? In the README file in github or in the paper itself? Thank you
