### 2018-08-22 #909 @whedon
[REVIEW]: bor: An R Package for Transforming Behavioral Observation Records into Data Matrices
**Submitting author:** @davidnsousa (<a href="http://orcid.org/0000-0001-7277-6447">David Sousa</a>)
**Repository:** <a href="https://github.com/davidnsousa/bor" target ="_blank">https://github.com/davidnsousa/bor</a>
**Version:** v0.1.0
