### 2018-08-22 #839 @ViniciusBRodrigues
@davidnsousa we [will proceed here](https://github.com/openjournals/joss-reviews/issues/909)
