### 2018-08-23 #909 @ViniciusBRodrigues
Hi @davidsousa, almost done!

There are few small stuffs to add in your repository to complete:
  * Automated test
  * Community guidelines (how people can contribute and report issues)

In pratical words: you need to create a folder with a R script for tests, like [here](https://github.com/williamsbenjamin/blendR). Also you can see how do it better [here](https://www.r-bloggers.com/unit-testing-with-r); Please, add the contact and how to contribute with the package in the README.md (that is, how the users can get help? email? Github? How to contribute?).

Best,
