### 2018-08-27 f4a2542
added automated test and contact
