+++ README.Rmd
@@ -29,9 +29,7 @@ more suitable for data analysis
 
 * `ex_field_data` a data frame containing an example of raw focal observations' data; data frames passed to `dtable()` function should have this structure.
 
-
-Read `bor` help file (`?bor`) to learn more about this package.
-
+For more information about this package the user can read the R documentation for `bor` (click the "rdocumentation" badge at the top of the README file). To learn more details about each `bor` object the user can type `?dtable`, `?countb` and `ex_field_data` in his R environment.
 
 ## Installation
 
@@ -108,3 +106,7 @@ Below we leave **bsep**, **asep**, **missing** and **noc**  `dtable()`'s argumen
 `data2` is a list of lists: one for each observer (in this example ``data2$`1` `` and ``data2$`2` ``). Inside each list there is one asymmetric interaction matrix per social interaction recorded (e.g., ``data2$`1`$`-` ``). Cells in these matrices provide counts on the number of times a specific type of social interaction was initiated by the row subject and directed to the column subject.
 
 See `?countb` for further details on the `countb()` function.
+
+## Contact
+
+For more help or any further questions about this package the user can e-mail the maintainer: davidnsousa@gmail.com. To contribute to this package the user can fork the package github repository and propose changes.
