### 2018-08-27 f35e516
changed details in README
