### 2018-08-27 3afef89
changed details in README
