### 2018-08-27 980e9a0
small fixes
