### 2018-08-28 #909 @davidnsousa
hello @karthik! The software is already archived in zenodo. Here is the DOI ![image](https://user-images.githubusercontent.com/5941390/44738244-3734a180-aaec-11e8-9adf-40e2de6e5b6d.png)

Thank you
