### 2018-08-28 #909 @karthik
Congrats @davidnsousa, your paper has been accepted. And thank you @ViniciusBRodrigues for the fast and timely review!  🎉

@arfon over to you for final processing.
