### 2018-08-29 #909 @davidnsousa
@arfon thank you! There are just two things that I have just notice. 1) the first reference in the text (Sousa & Daniel, 2018) should point to the DOI of the package in References but the DOI does not appear. Can I still remove this reference maybe it is not necessary. 2) The group of references in the first paragraph 1974, 1998 and 1993 does not appear to be in the right format in the text. Sorry that I only noticed this now. Is it still possible to correct these issues? Thank you very much for your help.

Best,
