### 2018-08-29 #909 @arfon
@davidnsousa - I suspect this is just the APA style for `@manual` etc. based on your entries in the bibtex: https://github.com/davidnsousa/bor/blob/master/paper.bib

Feel free to experiment with modifications to your bibtex here (in `master` on your repository) and use @whedon to rebuild the paper with `@whedon generate pdf`
