### 2018-08-29 #909 @davidnsousa
@arfon thank you! These issues are now corrected!
